refactor: StackNavigator에 type 추가
