+++ src/navigators/StackNavigator.tsx
@@ -1,9 +1,10 @@
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 
-import {StackMenu} from 'constants/menu';
+import {StackMenu} from 'navigators/constants';
+import {StackParamList} from 'navigators/types';
 import Home from 'screens/Home';
 
-const Stack = createNativeStackNavigator();
+const Stack = createNativeStackNavigator<StackParamList>();
 
 const StackNavigator = () => {
   return (
